refactor: adjust import path and change ErrorCode
